chore: move serializer signal impl to separate file
