+++ packages/qwik/src/core/signal/signal-api.ts
@@ -2,8 +2,10 @@ import type { QRLInternal } from '../shared/qrl/qrl-class';
 import type { QRL } from '../shared/qrl/qrl.public';
 import { SignalImpl } from './impl/signal-impl';
 import { ComputedSignalImpl } from './impl/computed-signal-impl';
-import { SerializerSignalImpl, throwIfQRLNotResolved, type SerializerArg } from './signal';
+import { throwIfQRLNotResolved } from './signal';
 import type { Signal } from './signal.public';
+import type { SerializerArg } from './types';
+import { SerializerSignalImpl } from './impl/serializer-signal-impl';
 
 /** @internal */
 export const createSignal = <T>(value?: T): Signal<T> => {
