+++ packages/qwik/src/core/signal/types.ts
@@ -6,6 +6,7 @@ import type { ReadonlySignal } from './signal.public';
 import type { TargetType } from './store';
 import type { SignalImpl } from './impl/signal-impl';
 import type { QRLInternal } from '../shared/qrl/qrl-class';
+import type { SerializerSymbol } from '../shared/utils/serialize-utils';
 
 export interface InternalReadonlySignal<T = unknown> extends ReadonlySignal<T> {
   readonly untrackedValue: T;
@@ -94,3 +95,53 @@ export const enum EffectProperty {
   COMPONENT = ':',
   VNODE = '.',
 }
+
+/** @public */
+export type SerializerArgObject<T, S> = {
+  /**
+   * This will be called with initial or serialized data to reconstruct an object. If no
+   * `initialData` is provided, it will be called with `undefined`.
+   *
+   * This must not return a Promise.
+   */
+  deserialize: (data: Awaited<S>) => T;
+  /** The initial value to use when deserializing. */
+  initial?: S | undefined;
+  /**
+   * This will be called with the object to get the serialized data. You can return a Promise if you
+   * need to do async work.
+   *
+   * The result may be anything that Qwik can serialize.
+   *
+   * If you do not provide it, the object will be serialized as `undefined`. However, if the object
+   * has a `[SerializerSymbol]` property, that will be used as the serializer instead.
+   */
+  serialize?: (obj: T) => S;
+};
+
+/**
+ * Serialize and deserialize custom objects.
+ *
+ * If you need to use scoped state, you can pass a function instead of an object. The function will
+ * be called with the current value, and you can return a new value.
+ *
+ * @public
+ */
+export type SerializerArg<T, S> =
+  | SerializerArgObject<T, S>
+  | (() => SerializerArgObject<T, S> & {
+      /**
+       * This gets called when reactive state used during `deserialize` changes. You may mutate the
+       * current object, or return a new object.
+       *
+       * If it returns a value, that will be used as the new value, and listeners will be triggered.
+       * If no change happened, don't return anything.
+       *
+       * If you mutate the current object, you must return it so that it will trigger listeners.
+       */
+      update?: (current: T) => T | void;
+    });
+
+export type CustomSerializable<T extends { [SerializerSymbol]: (obj: any) => any }, S> = {
+  [SerializerSymbol]: (obj: T) => S;
+};
