+++ packages/qwik/src/core/use/use-serializer.ts
@@ -1,7 +1,8 @@
 import { implicit$FirstArg } from '../shared/qrl/implicit_dollar';
 import type { QRL } from '../shared/qrl/qrl.public';
 import type { ComputedSignalImpl } from '../signal/impl/computed-signal-impl';
-import { SerializerSignalImpl, type SerializerArg } from '../signal/signal';
+import { SerializerSignalImpl } from '../signal/impl/serializer-signal-impl';
+import type { SerializerArg } from '../signal/types';
 import type { createSerializer$ } from '../signal/signal.public';
 import { useComputedCommon } from './use-computed';
